--- pkg/registry/interface/factory.go
@@ -1,13 +1,0 @@
-package factory
-
-type Registry interface {
-	Register(ctx context.Context, name string, serviceInstance *ServiceInstance) error
-	Unregister(ctx context.Context, name string) error
-	UnregisterAllService(ctx context.Context) error
-}
-
-type Discovery interface {
-	GetServices() []*ServiceInstance
-	WatchService() chan *ServiceInstance
-	Close()
-}
