+++ pkg/registry/registery.go
@@ -19,14 +19,6 @@ package registry
 
 import (
 	"context"
-	"fmt"
-	"github.com/arana-db/arana/pkg/registry/etcd"
-	"github.com/arana-db/arana/pkg/registry/store"
-	"github.com/docker/distribution/registry"
-	gostnet "github.com/dubbogo/gost/net"
-)
-
-import (
 	"github.com/arana-db/arana/pkg/config"
 )
 
@@ -42,57 +34,14 @@ type ServiceInstance struct {
 	Endpoints []*config.Listener
 }
 
-// DoRegistry register the service
-func DoRegistry(ctx context.Context, registryInstance Registry, name string, listeners []*config.Listener) error {
-	serviceInstance := &ServiceInstance{Name: name}
-	for _, listener := range listeners {
-		serviceInstance.Endpoints = append(serviceInstance.Endpoints, listener)
-	}
-
-	return registryInstance.Register(ctx, name, serviceInstance)
+type Registry interface {
+	Register(ctx context.Context, name string, serviceInstance *ServiceInstance) error
+	Unregister(ctx context.Context, name string) error
+	UnregisterAllService(ctx context.Context) error
 }
 
-func Init(registryConf *config.Registry) (registry.Registry, error) {
-	var serviceRegistry registry.Registry
-	var err error
-	switch registryConf.Name {
-	case store.ETCD:
-		serviceRegistry, err = initEtcd(registryConf)
-	case store.NACOS:
-	default:
-		err = errors.Errorf("Service registry not support store:%s", registryConf.Name)
-	}
-	if err != nil {
-		err = errors.Wrap(err, "init service registry err:%v")
-		log.Fatal(err.Error())
-		return nil, err
-	}
-	return serviceRegistry, nil
-}
-
-func initEtcd(registryConf *config.Registry) (registry.Registry, error) {
-	etcdAddr, ok := registryConf.Options["endpoints"].(string)
-	if !ok {
-		return nil, fmt.Errorf("service registry init etcd error because get endpoints of options :%v", registryConf.Options)
-	}
-
-	serverAddr, err := gostnet.GetLocalIP()
-	if !ok {
-		return nil, fmt.Errorf("service registry init etcd error because get local host err:%v", err)
-	}
-
-	rootPath, ok := registryConf.Options["root_path"].(string)
-	if !ok {
-		return nil, fmt.Errorf("service registry init etcd error because get root_path of options :%v", registryConf.Options)
-	}
-
-	serviceRegistry, err := etcd.NewEtcdV3Registry(serverAddr, rootPath, []string{etcdAddr}, nil)
-	if err != nil {
-		return nil, fmt.Errorf("service registry init etcd error because err: :%v", err)
-	}
-	return serviceRegistry, nil
-}
-
-func initNacos(registryConf *config.Registry) {
-
+type Discovery interface {
+	GetServices() []*ServiceInstance
+	WatchService() chan *ServiceInstance
+	Close()
 }
